+++ src/lib/AppRegistry.tsx
@@ -10,6 +10,8 @@ import {
   ArtistRenderer,
   BidderFlowRendererProps,
   BidFlowRenderer,
+  CityFairListRenderer,
+  CitySectionListRenderer,
   ConversationRenderer,
   FairRenderer,
   GeneRenderer,
@@ -36,6 +38,7 @@ import {
 import FavoritesScene from "./Scenes/Favorites"
 import HomeScene from "./Scenes/Home"
 import { MapContainer } from "./Scenes/Map"
+import { BucketKey } from "./Scenes/Map/Bucket"
 import renderWithLoadProgress from "./utils/renderWithLoadProgress"
 import { Schema, screenTrack as track } from "./utils/track"
 
@@ -185,6 +188,29 @@ const ShowMoreInfo: React.SFC<ShowMoreInfoProps> = ({ showID }) => {
   return <ShowMoreInfoRenderer showID={showID} render={renderWithLoadProgress(Containers.ShowMoreInfo, { showID })} />
 }
 
+interface CityFairListProps {
+  citySlug: string
+}
+const CityFairList: React.SFC<CityFairListProps> = ({ citySlug }) => {
+  return (
+    <CityFairListRenderer citySlug={citySlug} render={renderWithLoadProgress(Containers.CityFairList, { citySlug })} />
+  )
+}
+
+interface CitySectionListProps {
+  citySlug: string
+  section: BucketKey
+}
+const CitySectionList: React.SFC<CitySectionListProps> = ({ citySlug, section }) => {
+  return (
+    <CitySectionListRenderer
+      citySlug={citySlug}
+      section={section}
+      render={renderWithLoadProgress(Containers.CitySectionList, { citySlug, section })}
+    />
+  )
+}
+
 interface FairBoothProps {
   fairBoothID: string
 }
@@ -259,3 +285,5 @@ AppRegistry.registerComponent("ShowMoreInfo", () => ShowMoreInfo)
 AppRegistry.registerComponent("Map", () => MapContainer)
 AppRegistry.registerComponent("City", () => CityView)
 AppRegistry.registerComponent("CityPicker", () => CityPicker)
+AppRegistry.registerComponent("CityFairList", () => CityFairList)
+AppRegistry.registerComponent("CitySectionList", () => CitySectionList)